requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
python-dotenv>=1.0.0
aiohttp>=3.10.0
aiofiles>=24.1.0
rich>=13.9.0
typer>=0.13.0
sqlalchemy>=2.0.0
alembic>=1.13.0
httpx>=0.27.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
